Removed the redundant file
