+++ src/databricks/labs/remorph/reconcile/exception.py
@@ -1,11 +1,2 @@
 class ColumnMismatchException(Exception):
     """Raise the error when there is a mismatch in source and target column names"""
-
-
-def raise_column_mismatch_exception(msg: str, source_missing: list[str], target_missing: list[str]) -> Exception:
-    error_msg = (
-        f"{msg}\n"
-        f"columns missing in source: {','.join(source_missing) if source_missing else None}\n"
-        f"columns missing in target: {','.join(target_missing) if target_missing else None}\n"
-    )
-    return ColumnMismatchException(error_msg)
